requirements.txt
pygame
python-chess
torch
numpy